Move distributed in tensor/
