+++ crates/burn-tensor/src/lib.rs
@@ -56,7 +56,4 @@ pub use device::*;
 #[cfg(feature = "server")]
 pub mod server;
 
-#[cfg(feature = "distributed")]
-pub mod distributed;
-
 pub(crate) use burn_backend::TensorPrimitive;
