+++ crates/burn-tensor/src/tensor/mod.rs
@@ -46,6 +46,9 @@ pub mod ops {
 /// Tensor quantization module.
 pub mod quantization;
 
+#[cfg(feature = "distributed")]
+pub mod distributed;
+
 #[cfg(feature = "std")]
 pub use report::*;
