+++ crates/burn-tensor/src/tensor/module.rs
@@ -1,9 +1,5 @@
 use burn_backend::ops::ModuleOps;
-#[cfg(feature = "distributed")]
-use burn_backend::{DeviceOps, distributed::DistributedBackend};
 use burn_dispatch::Dispatch;
-#[cfg(feature = "distributed")]
-use burn_std::distributed::ReduceOperation;
 
 use crate::{
     Bool, Int, Tensor, check,
@@ -13,8 +9,6 @@ use crate::{
         InterpolateOptions, PadMode, PaddedConvOptions, UnfoldOptions,
     },
 };
-#[cfg(feature = "distributed")]
-use crate::{Device, distributed::CollectiveTensor};
 
 /// Computes the [CTC loss](burn_backend::ops::ModuleOps::ctc_loss).
 ///
@@ -642,25 +636,4 @@ fn layer_norm_impl(
         beta.map(|b| b.into_float()),
         epsilon,
     ))
-}
-
-/// Performs an all_reduce operation on the input tensor.
-///
-/// # Arguments
-/// - `input`: The input tensor.
-/// - `op`: The aggregation operation.
-/// - `device_ids`: The list of all devices with which to `all_reduce`
-///
-/// # Returns
-/// A [CollectiveTensor] containing the handle of the result.
-#[cfg(feature = "distributed")]
-pub fn all_reduce<const D: usize>(
-    input: Tensor<D>,
-    op: ReduceOperation,
-    device_ids: Vec<Device>,
-) -> CollectiveTensor<D> {
-    let device_ids = device_ids.iter().map(|d| d.as_dispatch().id()).collect();
-    let collective = Dispatch::all_reduce(input.primitive.into_float(), op, device_ids);
-    // Safety: we call `assume_resolved` only to wrap it in `burn_tensor`'s [CollectiveTensor].
-    CollectiveTensor::new(unsafe { collective.assume_resolved() })
 }
